+++ src/components/posts/posts/Tabs/share-tab.jsx
@@ -1,0 +1,11 @@
+import * as S from '../../../../styles/share-tab';
+
+const ShareTab = () => {
+    return(
+        <div>
+
+        </div>
+    );
+}
+
+export default ShareTab;
