+++ src/styles/share-tab.js
@@ -1,0 +1,2 @@
+import styled from "styled-components";
+
